refactor(context-manager): split offload.ts into per-strategy modules
Splits the monolithic offload.ts (~830 lines) into a directory:
- base.ts: BaseOffloadStrategy, EmergencyTruncateStrategy, shared helpers
- drop.ts: DropStrategy
- truncate.ts: TruncateStrategy
- summarize.ts: SummarizeStrategy
- index.ts: Offload namespace and re-exports
