--- strands-ts/src/context-manager/strategies/offload.ts
@@ -1,828 +1,0 @@
-/**
- * Builder API for offload strategies.
- *
- * Offload strategies reduce content in the context window.
- * The builder composes a target, a reduction method (truncate, drop, or summarize),
- * and optional conditions into a strategy that implements `ContextStrategy`.
- *
- * Conditions determine granularity:
- * - `threshold` only → per-block, eager (act on individual blocks above this size on message arrival)
- * - `utilization` only → message-level batch (remove/summarize oldest messages when utilization exceeded)
- * - Both → message-level, but only targets messages containing blocks over the threshold
- *
- * @internal
- */
-
-import { logger } from '../../logging/logger.js'
-import { MessageAddedEvent } from '../../hooks/events.js'
-import { Message, TextBlock, ToolResultBlock } from '../../types/messages.js'
-import type { ContentBlock } from '../../types/messages.js'
-import type { Model } from '../../models/model.js'
-import type { LocalAgent } from '../../types/agent.js'
-import type { ContextStrategy, ContextState } from '../types.js'
-import { truncateToolResultBlock, truncateTextBlock, type TruncateConfig } from './methods/truncate.js'
-import {
-  flattenMessagesToContent,
-  SUMMARIZED_PREFIX,
-  summarizeContent,
-  toolResultToContentBlocks,
-  type SummarizeConfig,
-} from './methods/summarize.js'
-
-export const DROPPED_MARKER = '[Dropped]'
-
-/**
- * Target for offload operations. This union is intentionally extensible — new
- * string-literal members can be added freely as new content categories emerge.
- *
- * - `"toolResults"` — all successful tool result blocks
- * - `"toolResultErrors"` — all failed tool result blocks
- * - `"assistantText"` — text blocks in assistant messages
- * - `"userText"` — text blocks in user messages (excluding tool results)
- * - `string[]` — tool results from specific tools, namespaced with `tool::` (e.g. `['tool::bash']`); prefix with `!` to exclude
- * - `"*"` — all content in the context window (tool results + text blocks)
- */
-export type OffloadTarget = '*' | 'toolResults' | 'toolResultErrors' | 'assistantText' | 'userText' | string[]
-
-/**
- * Conditions that determine when an offload strategy fires.
- *
- * Granularity is determined by which conditions are set:
- * - `threshold` only → per-block, eager (act on each block above this size on message arrival)
- * - `utilization` only → message-level (remove/summarize oldest messages when utilization exceeded)
- * - Both → message-level, targeting only messages with blocks over the threshold
- *
- * When multiple strategies target the same content, they don't conflict — strategies
- * run as an ordered pipeline, and once an earlier strategy shrinks a block, it falls
- * below the next strategy's threshold and gets skipped automatically.
- */
-export interface OffloadConditions {
-  /** Token threshold above which individual blocks are offloaded. */
-  threshold?: number
-
-  /** Context utilization ratio (0-1+) above which the strategy fires. */
-  utilization?: number
-
-  /** Number of most recent matching messages to leave untouched. */
-  preserveRecent?: number
-}
-
-/**
- * Intermediate builder result that allows chaining `.when()` conditions.
- * Also implements `ContextStrategy` directly so it can be used without `.when()`.
- */
-export interface OffloadStrategyBuilder extends ContextStrategy {
-  /** Add conditions that determine when this strategy fires. */
-  when(conditions: OffloadConditions): ContextStrategy
-}
-
-// --- Shared helpers ---
-
-function finiteOrUndefined(value: number | undefined): number | undefined {
-  return typeof value === 'number' && Number.isFinite(value) ? Math.max(0, value) : undefined
-}
-
-/**
- * Builds a toolUseId → toolName map from all assistant messages in the conversation.
- * Called once per apply() to avoid O(messages × toolResults) scanning.
- */
-function buildToolNameMap(messages: Message[]): Map<string, string> {
-  const map = new Map<string, string>()
-  for (const message of messages) {
-    if (message.role !== 'assistant') continue
-    for (const block of message.content) {
-      if ('toolUseId' in block && 'name' in block) {
-        map.set((block as { toolUseId: string }).toolUseId, (block as { name: string }).name)
-      }
-    }
-  }
-  return map
-}
-
-// --- Base strategy class ---
-
-/** Shared offload logic: target routing, eager hooks, preserveRecent. */
-abstract class BaseOffloadStrategy implements ContextStrategy {
-  abstract readonly name: string
-
-  protected readonly _target: OffloadTarget | undefined
-  protected readonly _threshold: number | undefined
-  protected readonly _utilization: number | undefined
-  protected readonly _preserveRecent: number
-  protected readonly _removalRatio: number = 0.3
-  protected readonly _includeFilter: Set<string> | undefined
-  protected readonly _excludeFilter: Set<string> | undefined
-
-  constructor(target?: OffloadTarget, conditions?: OffloadConditions) {
-    if (Array.isArray(target) && target.length === 0) {
-      throw new Error('Empty array target matches nothing — provide at least one target')
-    }
-
-    this._target = target
-    this._threshold = finiteOrUndefined(conditions?.threshold)
-    this._utilization = finiteOrUndefined(conditions?.utilization)
-    this._preserveRecent = Math.floor(finiteOrUndefined(conditions?.preserveRecent) ?? 0)
-
-    const resolved = resolveToolFilter(target)
-    this._includeFilter = resolved.include
-    this._excludeFilter = resolved.exclude
-  }
-
-  /** Whether this strategy operates at message-level (batch) vs per-block. */
-  protected get _isMessageLevel(): boolean {
-    return this._utilization !== undefined
-  }
-
-  init(agent: LocalAgent): void {
-    if (this._isMessageLevel) return
-    if (this._preserveRecent > 0) return
-    agent.addHook(MessageAddedEvent, async (event) => {
-      const messages = event.agent.messages
-      const toolNameMap = buildToolNameMap(messages)
-      await this._transformBlocks(event.message, messages, toolNameMap, event.agent)
-    })
-  }
-
-  async apply(context: ContextState): Promise<boolean> {
-    if (this._isMessageLevel) {
-      if (context.utilization < this._utilization!) return false
-      return this._applyPerMessage(context)
-    }
-
-    return this._applyPerBlock(context)
-  }
-
-  /** Per-block execution: walk each message, transform individual blocks above threshold. */
-  private async _applyPerBlock(context: ContextState): Promise<boolean> {
-    const { messages, agent } = context
-    const toolNameMap = buildToolNameMap(messages)
-    const eligible =
-      this._preserveRecent > 0
-        ? getOldestMatches(
-            messages,
-            this._target,
-            this._preserveRecent,
-            toolNameMap,
-            this._includeFilter,
-            this._excludeFilter
-          )
-        : messages
-    let acted = false
-
-    for (const message of eligible) {
-      if (await this._transformBlocks(message, messages, toolNameMap, agent)) {
-        acted = true
-      }
-    }
-
-    return acted
-  }
-
-  /** Message-level execution: remove oldest 30% of eligible messages with pair safety. */
-  protected async _applyPerMessage(context: ContextState): Promise<boolean> {
-    const { messages } = context
-    if (messages.length <= 1) return false
-
-    const eligible = await this._getEligibleMessages(context)
-    if (eligible.length === 0) return false
-
-    const targetRemoval = Math.max(1, Math.floor(eligible.length * this._removalRatio))
-    const toRemove = eligible.slice(0, targetRemoval)
-
-    const insertIndex = Math.max(1, messages.indexOf(toRemove[0]!))
-    const removed = spliceWithPairs(messages, toRemove)
-    if (removed === 0) return false
-
-    const marker = this._makeRemovalMarker(removed)
-    if (marker) {
-      messages.splice(
-        Math.min(insertIndex, messages.length),
-        0,
-        new Message({ role: 'user', content: [new TextBlock(marker)] })
-      )
-    }
-
-    repairAlternation(messages)
-    return true
-  }
-
-  /** Override to insert a marker when messages are removed. Return null for no marker. */
-  protected _makeRemovalMarker(_count: number): string | null {
-    return null
-  }
-
-  /** Process eligible blocks in a message. */
-  protected async _transformBlocks(
-    message: Message,
-    messages: Message[],
-    toolNameMap: Map<string, string>,
-    agent: LocalAgent
-  ): Promise<boolean> {
-    const effectiveThreshold = this._threshold ?? 0
-    let acted = false
-    for (let blockIndex = 0; blockIndex < message.content.length; blockIndex++) {
-      const block = message.content[blockIndex]!
-
-      if (block instanceof TextBlock) {
-        if (!targetMatchesMessage(this._target, message)) continue
-      } else if (block instanceof ToolResultBlock) {
-        if (
-          this._target !== undefined &&
-          !toolMatchesTarget(block, this._target, toolNameMap, this._includeFilter, this._excludeFilter)
-        )
-          continue
-      } else {
-        continue
-      }
-
-      const tokens = await agent.model.countTokens([new Message({ role: message.role, content: [block] })])
-      if (tokens <= effectiveThreshold) continue
-
-      const replacement = await this._replaceBlock(block, tokens, message, agent)
-      if (replacement && replacement !== block) {
-        ;(message.content as unknown[])[blockIndex] = replacement
-        acted = true
-      }
-    }
-    return acted
-  }
-
-  /** Collect eligible messages for message-level operations, respecting preserveRecent, head-pin, and threshold. */
-  protected async _getEligibleMessages(context: ContextState): Promise<Message[]> {
-    const { messages, agent } = context
-    const toolNameMap = buildToolNameMap(messages)
-
-    let candidates: Message[]
-    if (this._preserveRecent > 0) {
-      candidates = getOldestMatches(
-        messages,
-        this._target,
-        this._preserveRecent,
-        toolNameMap,
-        this._includeFilter,
-        this._excludeFilter
-      ).filter((message) => messages.indexOf(message) > 0)
-    } else {
-      candidates = messages.filter(
-        (message, index) =>
-          index > 0 &&
-          messageMatchesTarget(message, this._target, toolNameMap, this._includeFilter, this._excludeFilter)
-      )
-    }
-
-    if (this._threshold === undefined) return candidates
-
-    const eligible: Message[] = []
-    for (const message of candidates) {
-      if (await this._messageHasBlockOverThreshold(message, toolNameMap, agent)) {
-        eligible.push(message)
-      }
-    }
-    return eligible
-  }
-
-  /** Check if a message contains any target-matching block over the threshold. */
-  private async _messageHasBlockOverThreshold(
-    message: Message,
-    toolNameMap: Map<string, string>,
-    agent: LocalAgent
-  ): Promise<boolean> {
-    for (const block of message.content) {
-      if (block instanceof TextBlock) {
-        if (!targetMatchesMessage(this._target, message)) continue
-      } else if (block instanceof ToolResultBlock) {
-        if (
-          this._target !== undefined &&
-          !toolMatchesTarget(block, this._target, toolNameMap, this._includeFilter, this._excludeFilter)
-        )
-          continue
-      } else {
-        continue
-      }
-
-      const tokens = await agent.model.countTokens([new Message({ role: message.role, content: [block] })])
-      if (tokens > this._threshold!) return true
-    }
-    return false
-  }
-
-  /** Transform a block. Return the replacement, or null to skip. */
-  protected abstract _replaceBlock(
-    block: TextBlock | ToolResultBlock,
-    tokens: number,
-    message: Message,
-    agent: LocalAgent
-  ): Promise<ContentBlock | null>
-}
-
-// --- Drop strategy ---
-
-class DropStrategy extends BaseOffloadStrategy {
-  readonly name = 'offload:drop'
-
-  when(conditions: OffloadConditions): ContextStrategy {
-    return new DropStrategy(this._target, conditions)
-  }
-
-  protected override _makeRemovalMarker(count: number): string {
-    return `[Dropped: ${count} ${count === 1 ? 'message' : 'messages'}]`
-  }
-
-  protected async _replaceBlock(
-    block: TextBlock | ToolResultBlock,
-    _tokens: number,
-    message: Message,
-    _agent: LocalAgent
-  ): Promise<ContentBlock | null> {
-    if (block instanceof ToolResultBlock) {
-      logger.debug(`toolUseId=<${block.toolUseId}> | dropped tool result from context window`)
-      return new ToolResultBlock({
-        toolUseId: block.toolUseId,
-        status: block.status,
-        content: [new TextBlock(DROPPED_MARKER)],
-      })
-    }
-    logger.debug(`trackingId=<${message.trackingId}> | dropped text block from context window`)
-    return new TextBlock(DROPPED_MARKER)
-  }
-}
-
-// --- Truncate strategy ---
-
-class TruncateStrategy extends BaseOffloadStrategy {
-  readonly name = 'offload:truncate'
-
-  private readonly _truncateConfig: TruncateConfig
-
-  constructor(target?: OffloadTarget, config?: TruncateConfig, conditions?: OffloadConditions) {
-    super(target, conditions)
-    this._truncateConfig = config ?? {}
-
-    const previewTokens =
-      typeof this._truncateConfig.previewTokens === 'number' && Number.isFinite(this._truncateConfig.previewTokens)
-        ? this._truncateConfig.previewTokens
-        : 1000
-    if (
-      conditions?.threshold !== undefined &&
-      Number.isFinite(conditions.threshold) &&
-      conditions.threshold <= previewTokens
-    ) {
-      throw new Error(
-        `threshold (${conditions.threshold}) must be greater than previewTokens (${previewTokens}) to ensure truncation converges`
-      )
-    }
-  }
-
-  when(conditions: OffloadConditions): ContextStrategy {
-    return new TruncateStrategy(this._target, this._truncateConfig, conditions)
-  }
-
-  protected override _makeRemovalMarker(count: number): string {
-    return `[... ${count} ${count === 1 ? 'message' : 'messages'} elided ...]`
-  }
-
-  protected override async _applyPerMessage(context: ContextState): Promise<boolean> {
-    const { messages } = context
-    if (messages.length <= 1) return false
-
-    const eligible = await this._getEligibleMessages(context)
-    if (eligible.length === 0) return false
-
-    // Determine head/tail split based on config (default: favor tail — 30% head, 70% tail)
-    const previewMode = this._truncateConfig.preview ?? 'headTail'
-    const headShare = { head: 1, tail: 0, headTail: 0.3 }[previewMode]
-    const targetRemoval = Math.max(1, Math.floor(eligible.length * this._removalRatio))
-    const keepCount = eligible.length - targetRemoval
-
-    const headKeep = Math.floor(keepCount * headShare)
-    const tailKeep = keepCount - headKeep
-
-    // Split into head (keep), middle (remove), tail (keep)
-    const headMessages = eligible.slice(0, headKeep)
-    const middleMessages = eligible.slice(headKeep, eligible.length - (tailKeep || 0))
-
-    if (middleMessages.length === 0) return false
-
-    const insertIndex = headKeep > 0 ? messages.indexOf(headMessages[headKeep - 1]!) + 1 : 1
-    const removed = spliceWithPairs(messages, middleMessages)
-    if (removed === 0) return false
-
-    const marker = this._makeRemovalMarker(removed)
-    messages.splice(
-      Math.min(insertIndex, messages.length),
-      0,
-      new Message({ role: 'user', content: [new TextBlock(marker)] })
-    )
-
-    repairAlternation(messages)
-    return true
-  }
-
-  protected async _replaceBlock(
-    block: TextBlock | ToolResultBlock,
-    tokens: number,
-    message: Message,
-    _agent: LocalAgent
-  ): Promise<ContentBlock | null> {
-    if (block instanceof ToolResultBlock) {
-      logger.debug(`toolUseId=<${block.toolUseId}>, tokens=<${tokens}> | truncated tool result`)
-      return truncateToolResultBlock(block, this._truncateConfig)
-    }
-    logger.debug(`trackingId=<${message.trackingId}>, tokens=<${tokens}> | truncated text block`)
-    return truncateTextBlock(block, this._truncateConfig)
-  }
-}
-
-// --- Summarize strategy ---
-
-class SummarizeStrategy extends BaseOffloadStrategy {
-  readonly name = 'offload:summarize'
-
-  private readonly _config: SummarizeConfig
-
-  constructor(target?: OffloadTarget, config?: SummarizeConfig, conditions?: OffloadConditions) {
-    super(target, conditions)
-    this._config = config ?? {}
-  }
-
-  when(conditions: OffloadConditions): ContextStrategy {
-    return new SummarizeStrategy(this._target, this._config, conditions)
-  }
-
-  override init(agent: LocalAgent): void {
-    if (this._utilization !== undefined) return
-    super.init(agent)
-  }
-
-  override async apply(context: ContextState): Promise<boolean> {
-    if (!this._resolveModel(context.agent)) {
-      logger.warn('no model available for summarization')
-      return false
-    }
-    if (context.messages.length === 0) return false
-    return super.apply(context)
-  }
-
-  protected override async _applyPerMessage(context: ContextState): Promise<boolean> {
-    const model = this._resolveModel(context.agent)
-    if (!model) return false
-
-    const { messages } = context
-    if (messages.length <= 1) return false
-
-    const eligible = await this._getEligibleMessages(context)
-    if (eligible.length === 0) return false
-
-    const summarizeCount = Math.max(1, Math.floor(eligible.length * this._removalRatio))
-    const toSummarize = eligible.slice(0, summarizeCount)
-
-    // Expand to include paired messages so we don't orphan tool pairs
-    const safeSet = new Set<Message>()
-    for (const message of toSummarize) {
-      const index = messages.indexOf(message)
-      if (index === -1) continue
-      for (const removable of collectRemovableWithPair(messages, index)) {
-        safeSet.add(removable)
-      }
-    }
-    const safe = messages.filter((message) => safeSet.has(message))
-    if (safe.length === 0) return false
-
-    const contentBlocks = flattenMessagesToContent(safe)
-    const summary = await summarizeContent(contentBlocks, model, this._config)
-    if (!summary) return false
-
-    const totalTokens = await model.countTokens(safe)
-    const summaryMessage = new Message({
-      role: 'user',
-      content: [
-        new TextBlock(
-          `${SUMMARIZED_PREFIX} ${safe.length} messages, ~${totalTokens.toLocaleString()} tokens]\n\n${summary}`
-        ),
-      ],
-    })
-
-    // Record insertion point before removing (position of the first summarized message)
-    const insertIndex = Math.max(1, messages.indexOf(safe[0]!))
-
-    // Remove summarized messages individually (they may not be contiguous)
-    for (const message of safe) {
-      const index = messages.indexOf(message)
-      if (index !== -1) messages.splice(index, 1)
-    }
-
-    // Insert summary where the first summarized message was
-    const clampedInsert = Math.min(insertIndex, messages.length)
-    messages.splice(clampedInsert, 0, summaryMessage)
-
-    repairAlternation(messages)
-    logger.debug(`summarized=<${safe.length}>, tokens=<${totalTokens}> | batched summarization complete`)
-    return true
-  }
-
-  protected async _replaceBlock(
-    block: TextBlock | ToolResultBlock,
-    tokens: number,
-    message: Message,
-    agent: LocalAgent
-  ): Promise<ContentBlock | null> {
-    const model = this._resolveModel(agent)
-    if (!model) return null
-
-    if (block instanceof ToolResultBlock) {
-      const summary = await summarizeContent(toolResultToContentBlocks(block.content), model, this._config)
-      if (!summary) return null
-
-      logger.debug(`toolUseId=<${block.toolUseId}>, tokens=<${tokens}> | summarized tool result`)
-      return new ToolResultBlock({
-        toolUseId: block.toolUseId,
-        status: block.status,
-        content: [new TextBlock(`${SUMMARIZED_PREFIX} ~${tokens.toLocaleString()} tokens]\n\n${summary}`)],
-      })
-    }
-
-    const summary = await summarizeContent([new TextBlock(block.text)], model, this._config)
-    if (!summary) return null
-
-    logger.debug(`trackingId=<${message.trackingId}>, tokens=<${tokens}> | summarized text block`)
-    return new TextBlock(`${SUMMARIZED_PREFIX} ~${tokens.toLocaleString()} tokens]\n\n${summary}`)
-  }
-
-  private _resolveModel(agent: LocalAgent): Model | undefined {
-    return this._config.model ?? agent.model
-  }
-}
-
-// --- preserveRecent helper ---
-
-/**
- * Returns target-matching messages excluding the N most recent matches.
- * First filters to only messages that match the target, then removes the last N from that set.
- */
-function getOldestMatches(
-  messages: Message[],
-  target: OffloadTarget | undefined,
-  count: number,
-  toolNameMap: Map<string, string>,
-  toolFilter: Set<string> | undefined,
-  excludeFilter: Set<string> | undefined
-): Message[] {
-  const matching = messages.filter((message) =>
-    messageMatchesTarget(message, target, toolNameMap, toolFilter, excludeFilter)
-  )
-  if (count >= matching.length) return []
-  return matching.slice(0, -count)
-}
-
-function toolMatchesTarget(
-  block: ToolResultBlock,
-  target: OffloadTarget,
-  toolNameMap: Map<string, string>,
-  toolFilter: Set<string> | undefined,
-  excludeFilter: Set<string> | undefined
-): boolean {
-  if (target === 'toolResults') return block.status === 'success'
-  if (target === 'toolResultErrors') return block.status === 'error'
-
-  const toolName = toolNameMap.get(block.toolUseId)
-  if (!toolName) return toolFilter === undefined && excludeFilter === undefined
-
-  if (excludeFilter) return !excludeFilter.has(toolName)
-  if (toolFilter) return toolFilter.has(toolName)
-
-  return true
-}
-
-function targetMatchesMessage(target: OffloadTarget | undefined, message: Message): boolean {
-  if (target === undefined || target === '*') return true
-  if (target === 'assistantText') return message.role === 'assistant'
-  if (target === 'userText') return message.role === 'user'
-  return false
-}
-
-function messageMatchesTarget(
-  message: Message,
-  target: OffloadTarget | undefined,
-  toolNameMap: Map<string, string>,
-  toolFilter: Set<string> | undefined,
-  excludeFilter: Set<string> | undefined
-): boolean {
-  if (targetMatchesMessage(target, message)) return true
-  if (!target) return false
-  if (target === 'assistantText' || target === 'userText') return false
-
-  // Tool result targets — must be a user message with a matching tool result
-  if (message.role !== 'user') return false
-  for (const block of message.content) {
-    if (block instanceof ToolResultBlock) {
-      if (toolMatchesTarget(block, target, toolNameMap, toolFilter, excludeFilter)) return true
-    }
-  }
-  return false
-}
-
-/**
- * Collects a message and its paired partner (if any) for safe removal.
- * If removing a message would orphan a tool-use/tool-result pair, includes the partner
- * so the pair is removed together. Skips messages[0] (head-pin).
- */
-function collectRemovableWithPair(messages: Message[], index: number): Message[] {
-  const message = messages[index]
-  if (!message) return []
-  if (index === 0) return []
-
-  const result: Message[] = [message]
-
-  const hasToolResult = message.content.some((block) => block.type === 'toolResultBlock')
-  if (hasToolResult) {
-    const prev = messages[index - 1]
-    if (prev && prev.content.some((block) => block.type === 'toolUseBlock')) {
-      if (index - 1 > 0) result.push(prev)
-      else return []
-    }
-  }
-
-  const hasToolUse = message.content.some((block) => block.type === 'toolUseBlock')
-  if (hasToolUse && index < messages.length - 1) {
-    const next = messages[index + 1]
-    if (next && next.content.some((block) => block.type === 'toolResultBlock')) {
-      result.push(next)
-    }
-  }
-
-  return result
-}
-
-/**
- * Collects removable messages (with their tool pairs), then splices them from the array.
- * Returns the number of messages removed.
- */
-function spliceWithPairs(messages: Message[], toRemove: Message[]): number {
-  const toSplice = new Set<Message>()
-  for (const message of toRemove) {
-    const index = messages.indexOf(message)
-    if (index === -1) continue
-    for (const removable of collectRemovableWithPair(messages, index)) {
-      toSplice.add(removable)
-    }
-  }
-
-  let removed = 0
-  for (const message of toSplice) {
-    const index = messages.indexOf(message)
-    if (index === -1) continue
-    messages.splice(index, 1)
-    removed++
-  }
-  return removed
-}
-
-// --- Builder ---
-
-/**
- * Offload strategy builder namespace.
- *
- * - `Offload.drop(target)` — drop matching content from context window entirely
- * - `Offload.truncate(target, config)` — replace with a preview
- * - `Offload.summarize(target, config)` — replace with LLM-generated summary
- */
-interface OffloadNamespace {
-  /** Drop matching content from context window entirely. */
-  drop(target: OffloadTarget): OffloadStrategyBuilder
-
-  /** Replace oversized content with a preview. */
-  truncate(target: OffloadTarget, config?: TruncateConfig): OffloadStrategyBuilder
-
-  /** Replace oversized content with an LLM-generated summary. */
-  summarize(target: OffloadTarget, config?: SummarizeConfig): OffloadStrategyBuilder
-}
-
-/**
- * Builder for offload strategies — reduces content in the context window.
- *
- * @example
- * ```typescript
- * // Per-block: truncate each result over 2500 tokens, eagerly
- * Offload.truncate("toolResults", { previewTokens: 750 }).when({ threshold: 1500 })
- * // Per-block: truncate specific tools
- * Offload.truncate(["tool::bash", "tool::read_file"]).when({ threshold: 2000 })
- * // Message-level: summarize oldest messages on overflow
- * Offload.summarize("*").when({ utilization: 1, preserveRecent: 4 })
- * // Per-block: drop errors over 500 tokens
- * Offload.drop("toolResultErrors").when({ threshold: 500 })
- * ```
- */
-export const Offload: OffloadNamespace = {
-  drop(target: OffloadTarget): OffloadStrategyBuilder {
-    return new DropStrategy(target)
-  },
-
-  truncate(target: OffloadTarget, config?: TruncateConfig): OffloadStrategyBuilder {
-    return new TruncateStrategy(target, config)
-  },
-
-  summarize(target: OffloadTarget, config?: SummarizeConfig): OffloadStrategyBuilder {
-    return new SummarizeStrategy(target, config)
-  },
-}
-
-// --- Emergency truncation strategy ---
-
-/**
- * Last-resort strategy that recomputes utilization and drops the oldest 20% of messages
- * when the context window is still overflowing after all user-configured strategies have run.
- * Appended internally by ContextManager — not part of the public builder API.
- *
- * TODO: add a ContextManagerConfig flag to disable this for users who want full control.
- *
- * @internal
- */
-export class EmergencyTruncateStrategy extends BaseOffloadStrategy {
-  readonly name = 'offload:emergency-truncate'
-  protected override readonly _removalRatio = 0.2
-
-  constructor() {
-    super('*')
-  }
-
-  override init(): void {
-    // No eager hooks — this only fires on overflow
-  }
-
-  override async apply(context: ContextState): Promise<boolean> {
-    if (context.messages.length <= 3) return false
-    const tokens = await context.agent.model.countTokens(context.messages)
-    const utilization = context.agent.model.estimateUtilization(tokens)
-    if (utilization < 1.0) return false
-    return this._applyPerMessage({ ...context, utilization })
-  }
-
-  protected async _replaceBlock(): Promise<ContentBlock | null> {
-    return null
-  }
-}
-
-// --- Role alternation repair ---
-
-/**
- * Merges consecutive same-role messages to restore the user/assistant alternation
- * that Anthropic/Bedrock APIs require. Called after message-level operations that
- * may leave gaps. Constructs new Message objects for merges to avoid mutating
- * objects that may be referenced elsewhere (session storage, event payloads).
- */
-function repairAlternation(messages: Message[]): void {
-  let writeIndex = 0
-  for (let readIndex = 0; readIndex < messages.length; readIndex++) {
-    const current = messages[readIndex]!
-    if (writeIndex > 0 && messages[writeIndex - 1]!.role === current.role) {
-      const prev = messages[writeIndex - 1]!
-      messages[writeIndex - 1] = new Message({
-        role: prev.role,
-        content: [...prev.content, ...current.content],
-      })
-    } else {
-      messages[writeIndex] = current
-      writeIndex++
-    }
-  }
-  messages.length = writeIndex
-}
-
-// --- Helpers ---
-
-/**
- * Parses a string[] target into include/exclude filter sets.
- * Entries must be prefixed with `tool::` (e.g. `'tool::bash'`).
- * An additional `!` prefix excludes (e.g. `'!tool::bash'`).
- */
-function resolveToolFilter(target: OffloadTarget | undefined): { include?: Set<string>; exclude?: Set<string> } {
-  if (target === undefined || target === '*') return {}
-  if (typeof target === 'string') return {}
-  if (!Array.isArray(target)) return {}
-
-  const TOOL_PREFIX = 'tool::'
-  const includes: string[] = []
-  const excludes: string[] = []
-
-  for (const entry of target) {
-    if (entry.startsWith('!')) {
-      const name = entry.slice(1)
-      excludes.push(name.startsWith(TOOL_PREFIX) ? name.slice(TOOL_PREFIX.length) : name)
-    } else {
-      includes.push(entry.startsWith(TOOL_PREFIX) ? entry.slice(TOOL_PREFIX.length) : entry)
-    }
-  }
-
-  if (excludes.length > 0 && includes.length > 0) {
-    logger.warn('tool filter contains both includes and excludes, excludes will be ignored')
-    return { include: new Set(includes) }
-  }
-  if (excludes.length > 0) {
-    return { exclude: new Set(excludes) }
-  }
-  if (includes.length > 0) {
-    return { include: new Set(includes) }
-  }
-
-  return {}
-}
